refactor(room): move `room` to separate file
