+++ src/prep/prep.hh
@@ -14,31 +14,7 @@ enum EventType {
     PLAYER_STATE_CHANGE,
 };
 
-enum RoomStatus {
-    AWAITING_START,
-    IN_PROGRESS,
-    STOPPED,
-    ENDED,
-};
-
-struct Room;
 struct Event;
-
-struct Room {
-    uint32_t id;
-    std::string title;
-    std::tm *utcTimestampCreatedAt;
-    RoomStatus status;
-    std::vector<Player> players;
-
-    Room(
-      uint32_t id, std::string title, uint32_t utcTimestampCreatedAt, RoomStatus status, std::vector<Player> players);
-    Room(uint32_t id,
-         std::string title,
-         uint32_t utcTimestampCreatedAt,
-         RoomStatus status,
-         std::initializer_list<Player> players);
-};
 
 struct Event {
     std::string title;
